refactor: Major => UserMajor
