+++ src/entities/UserMajor.ts
